requirements.txt
# Core detection dependencies
torch>=2.0.0
torchvision>=0.15.0
ultralytics>=8.0.0  # YOLOv8
opencv-python>=4.8.0
numpy>=1.24.0
pillow>=10.0.0

# Configuration and data
pyyaml>=6.0

# Async and utilities
asyncio  # Built into Python
aiofiles>=23.0.0
python-dotenv>=1.0.0

# Logging and monitoring
loguru>=0.7.0

# Optional: Alert notifications (configure as needed)
# twilio>=8.0.0  # SMS alerts
# smtplib  # Built into Python - email alerts

# Development and testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.0.0
flake8>=6.0.0